LPD-62838 Addding validation to check if Object Entry status is equal to STATUS_IN_TRASH

+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/manager/v1_0/DefaultObjectEntryManagerImpl.java
@@ -20,6 +20,8 @@ import com.liferay.object.constants.ObjectRelationshipConstants;
 import com.liferay.object.entry.folder.subscription.util.ObjectEntryFolderSubscriptionUtil;
 import com.liferay.object.entry.util.ObjectEntryDTOConverterUtil;
 import com.liferay.object.exception.NoSuchObjectEntryException;
+import com.liferay.object.exception.ObjectEntryValuesException;
+import com.liferay.object.exception.ObjectEntryVersionStatusException;
 import com.liferay.object.field.attachment.AttachmentManager;
 import com.liferay.object.field.business.type.ObjectFieldBusinessType;
 import com.liferay.object.field.business.type.ObjectFieldBusinessTypeRegistry;
@@ -288,6 +290,13 @@ public class DefaultObjectEntryManagerImpl
 		ObjectEntry objectEntry = getObjectEntryByVersion(
 			dtoConverterContext, objectEntryId, version);
 
+		Status status = objectEntry.getStatus();
+
+		if (status.getCode() == WorkflowConstants.STATUS_IN_TRASH) {
+			throw new ObjectEntryVersionStatusException.
+				MustNotCopyObjectEntryVersionInTrash();
+		}
+
 		return _copyVersionedObjectEntry(
 			dtoConverterContext, objectDefinition, objectEntry);
 	}
@@ -306,6 +315,13 @@ public class DefaultObjectEntryManagerImpl
 		ObjectEntry objectEntry = getObjectEntryByVersion(
 			dtoConverterContext, externalReferenceCode, objectDefinition,
 			scopeKey, version);
+
+		Status status = objectEntry.getStatus();
+
+		if (status.getCode() == WorkflowConstants.STATUS_IN_TRASH) {
+			throw new ObjectEntryVersionStatusException.
+				MustNotCopyObjectEntryVersionInTrash();
+		}
 
 		return _copyVersionedObjectEntry(
 			dtoConverterContext, objectDefinition, objectEntry);
@@ -349,9 +365,18 @@ public class DefaultObjectEntryManagerImpl
 			throw new UnsupportedOperationException();
 		}
 
+		com.liferay.object.model.ObjectEntry serviceBuilderObjectEntry =
+			_objectEntryService.getObjectEntry(objectEntryId);
+
 		_checkObjectEntryObjectDefinitionId(
-			objectDefinition,
-			_objectEntryService.getObjectEntry(objectEntryId));
+			objectDefinition, serviceBuilderObjectEntry);
+
+		if (serviceBuilderObjectEntry.getStatus() ==
+				WorkflowConstants.STATUS_IN_TRASH) {
+
+			throw new ObjectEntryVersionStatusException.
+				MustNotDeleteObjectEntryVersionInTrash();
+		}
 
 		_objectEntryVersionService.deleteObjectEntryVersion(
 			objectEntryId, version);
@@ -374,6 +399,13 @@ public class DefaultObjectEntryManagerImpl
 
 		_checkObjectEntryObjectDefinitionId(
 			objectDefinition, serviceBuilderObjectEntry);
+
+		if (serviceBuilderObjectEntry.getStatus() ==
+				WorkflowConstants.STATUS_IN_TRASH) {
+
+			throw new ObjectEntryVersionStatusException.
+				MustNotDeleteObjectEntryVersionInTrash();
+		}
 
 		_objectEntryVersionService.deleteObjectEntryVersion(
 			serviceBuilderObjectEntry.getObjectEntryId(), version);
@@ -506,9 +538,19 @@ public class DefaultObjectEntryManagerImpl
 			ObjectDefinition objectDefinition, long objectEntryId, int version)
 		throws Exception {
 
+		com.liferay.object.model.ObjectEntry serviceBuilderObjectEntry =
+			_objectEntryService.getObjectEntry(objectEntryId);
+
+		if (serviceBuilderObjectEntry.getStatus() ==
+				WorkflowConstants.STATUS_IN_TRASH) {
+
+			throw new ObjectEntryVersionStatusException.
+				MustNotExpireObjectEntryVersionInTrash();
+		}
+
 		return _expireObjectEntryVersion(
-			dtoConverterContext, objectDefinition,
-			_objectEntryService.getObjectEntry(objectEntryId), version);
+			dtoConverterContext, objectDefinition, serviceBuilderObjectEntry,
+			version);
 	}
 
 	@Override
@@ -518,11 +560,20 @@ public class DefaultObjectEntryManagerImpl
 			String scopeKey, int version)
 		throws Exception {
 
-		return _expireObjectEntryVersion(
-			dtoConverterContext, objectDefinition,
+		com.liferay.object.model.ObjectEntry serviceBuilderObjectEntry =
 			_objectEntryService.getObjectEntry(
 				externalReferenceCode, getGroupId(objectDefinition, scopeKey),
-				objectDefinition.getObjectDefinitionId()),
+				objectDefinition.getObjectDefinitionId());
+
+		if (serviceBuilderObjectEntry.getStatus() ==
+				WorkflowConstants.STATUS_IN_TRASH) {
+
+			throw new ObjectEntryVersionStatusException.
+				MustNotExpireObjectEntryVersionInTrash();
+		}
+
+		return _expireObjectEntryVersion(
+			dtoConverterContext, objectDefinition, serviceBuilderObjectEntry,
 			version);
 	}
 
@@ -1128,10 +1179,18 @@ public class DefaultObjectEntryManagerImpl
 			ObjectDefinition objectDefinition, long objectEntryId, int version)
 		throws Exception {
 
+		ObjectEntry objectEntry = getObjectEntryByVersion(
+			dtoConverterContext, objectEntryId, version);
+
+		Status status = objectEntry.getStatus();
+
+		if (status.getCode() == WorkflowConstants.STATUS_IN_TRASH) {
+			throw new ObjectEntryVersionStatusException.
+				MustNotRestoreObjectEntryVersionInTrash();
+		}
+
 		return _restoreVersionedObjectEntry(
-			dtoConverterContext, objectDefinition,
-			getObjectEntryByVersion(
-				dtoConverterContext, objectEntryId, version));
+			dtoConverterContext, objectDefinition, objectEntry);
 	}
 
 	@Override
@@ -1141,11 +1200,19 @@ public class DefaultObjectEntryManagerImpl
 			String scopeKey, int version)
 		throws Exception {
 
+		ObjectEntry objectEntry = getObjectEntryByVersion(
+			dtoConverterContext, externalReferenceCode, objectDefinition,
+			scopeKey, version);
+
+		Status status = objectEntry.getStatus();
+
+		if (status.getCode() == WorkflowConstants.STATUS_IN_TRASH) {
+			throw new ObjectEntryVersionStatusException.
+				MustNotRestoreObjectEntryVersionInTrash();
+		}
+
 		return _restoreVersionedObjectEntry(
-			dtoConverterContext, objectDefinition,
-			getObjectEntryByVersion(
-				dtoConverterContext, externalReferenceCode, objectDefinition,
-				scopeKey, version));
+			dtoConverterContext, objectDefinition, objectEntry);
 	}
 
 	@Override
